requirements.txt
selenium==4.17.2
pandas==2.0.2
google-cloud-storage==2.12.0
webdriver_manager==4.0.1
PyYAML==6.0
pyarrow==11.0.0
fastparquet